+++ setup.cfg
@@ -25,5 +25,8 @@ strict_optional = True
 warn_redundant_casts = True
 warn_unused_ignores = True
 
+[mypy-pypiserver.bottle]
+follow_imports = skip
+
 [mypy-tests.*]
 disallow_untyped_decorators = False
